Bugfix: currency loss for multiplication and division of non-current-culture currencies
For instance if your current culture's currency is EUR and you multiply USDs, you get EUR as the result, due to the implicit conversion from decimal to Money:

USD 2 * 2 = EUR 4

This is circumvented by explicitly calling the constructor with the same currency as the left operand. The same goes for the division operator.